Remove WithdrawalHistoryProvider
