+++ src/Stratis.FederatedPeg.Features.FederationGateway/FederationGatewayFeature.cs
@@ -90,8 +90,6 @@ namespace Stratis.FederatedPeg.Features.FederationGateway
 
         private readonly IMaturedBlocksSyncManager maturedBlocksSyncManager;
 
-        private readonly IWithdrawalHistoryProvider withdrawalHistoryProvider;
-
         private readonly ILogger logger;
 
         public FederationGatewayFeature(
@@ -112,8 +110,7 @@ namespace Stratis.FederatedPeg.Features.FederationGateway
             IPartialTransactionRequester partialTransactionRequester,
             IFederationGatewayClient federationGatewayClient,
             MempoolManager mempoolManager,
-            IMaturedBlocksSyncManager maturedBlocksSyncManager,
-            IWithdrawalHistoryProvider withdrawalHistoryProvider)
+            IMaturedBlocksSyncManager maturedBlocksSyncManager)
         {
             this.loggerFactory = loggerFactory;
             this.signals = signals;
@@ -132,7 +129,6 @@ namespace Stratis.FederatedPeg.Features.FederationGateway
             this.federationGatewayClient = federationGatewayClient;
             this.mempoolManager = mempoolManager;
             this.maturedBlocksSyncManager = maturedBlocksSyncManager;
-            this.withdrawalHistoryProvider = withdrawalHistoryProvider;
 
             this.logger = loggerFactory.CreateLogger(this.GetType().FullName);
 
@@ -241,7 +237,7 @@ namespace Stratis.FederatedPeg.Features.FederationGateway
             }
 
             // Display recent withdrawals (if any).
-            List<WithdrawalModel> withdrawals = this.withdrawalHistoryProvider.GetHistory(5);
+            List<WithdrawalModel> withdrawals = this.federationWalletManager.GetHistory(5);
             if (withdrawals.Count > 0)
             {
                 benchLog.AppendLine("--- Recent Withdrawals ---");
@@ -327,7 +323,6 @@ namespace Stratis.FederatedPeg.Features.FederationGateway
                         services.AddSingleton<IPartialTransactionRequester, PartialTransactionRequester>();
                         services.AddSingleton<IFederationGatewayClient, FederationGatewayClient>();
                         services.AddSingleton<IMaturedBlocksSyncManager, MaturedBlocksSyncManager>();
-                        services.AddSingleton<IWithdrawalHistoryProvider, WithdrawalHistoryProvider>();
                     });
             });
             return fullNodeBuilder;
